backend/requirements.txt
fastapi==0.116.1
openai==1.97.0
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.1.1
